refactor : rowmapper 사용하여 수정
